pidfs: use maple tree
Hey,

Ok, I wanted to give this another try as I'd really like to rely on the
maple tree supporting ULONG_MAX when BITS_PER_LONG is 64 as it makes
things a lot simpler overall.

As Willy didn't want additional users relying on an external lock I made
it so that we don't have to and can just use the mtree lock.

However, I need an irq safe variant which is why I added support for
this into the maple tree.

This is pullable from
https://git.kernel.org/pub/scm/linux/kernel/git/vfs/vfs.git work.pidfs.maple_tree

Thanks!
Christian

To: Liam R. Howlett <Liam.Howlett@oracle.com>
To: Matthew Wilcox <willy@infradead.org>
Cc: linux-fsdevel@vger.kernel.org
Cc: maple-tree@lists.infradead.org

---
Changes in v3:
- EDITME: describe what is new in this series revision.
- EDITME: use bulletpoints and terse descriptions.
- Link to v2: https://lore.kernel.org/r/20241209-work-pidfs-maple_tree-v2-0-003dbf3bd96b@kernel.org



--- b4-submit-tracking ---
# This section is used internally by b4 prep for tracking purposes.
{
  "series": {
    "revision": 3,
    "change-id": "20241206-work-pidfs-maple_tree-b322ff5399b7",
    "prefixes": [
      "RFC"
    ],
    "base-branch": "vfs-6.14.pidfs",
    "history": {
      "v1": [
        "20241206-work-pidfs-maple_tree-v1-1-1cca6731b67f@kernel.org"
      ],
      "v2": [
        "20241209-work-pidfs-maple_tree-v2-0-003dbf3bd96b@kernel.org"
      ]
    }
  }
}