pidfs: switch inode number handling to rbtree
The new pid inode number allocation scheme was clever but I overlooked
a possible, even though unlikely, attack that can be used to trigger an
overflow on both 32bit and 64bit.

An inode number was constructed by allocation a 32bit number using the
idr_alloc_cyclic() infrastructure. When the idr wrapped around a
a 32bit counter was incremented. The 32bit counter served as the upper
32bits and the allocated idr number as the lower 32bit for the 64bit
inode number stored in pid->ino.

There's a few problems with this. Since the idr can only allocate up to
INT_MAX entries everytime a wraparound happens INT_MAX - 1 entries are
lost. If userspace densely populate the idr such that only two entries
remain, an entry < INT_MAX and the INT_MAX entry, it is possible to
overflow the wraparound counter. That is unlikely but the mere
possibility is pretty annoying and goes against the guarantees pidfs
wants to give.

Various approaches were considered. By using an xarray it would mean the
wraparound is pushed to UINT_MAX but that's not that much of an
improvement compared to the idr and it doesn't allow us to simplify
things. Alternatively, the maple_tree could be used. This would give us
ULONG_MAX on 64bit and thus would prevent any wraparound from occuring.
But erasing entries is non-trivial in contexts where we are under the
irq write lock of tasklist_lock. This can potentially be fixed in the
future.

But for now, switch pidfs to an rbtree for lookup of struct pid based on
inode numbers. This means we can lookup struct pid based on the 64 bit
inode number on both 64bit and 32bit. On 32bit we split the 64bit inode
number into the lower 32bit inode number and the upper bit 32bit into
the generation number.

Signed-off-by: Christian Brauner <brauner@kernel.org>