pidfs: ensure consistent ENOENT/ESRCH reporting
In a prior patch series we tried to cleanly differentiate between:

(1) The task has already been reaped.
(2) The caller requested a pidfd for a thread-group leader but the pid
actually references a struct pid that isn't used as a thread-group
leader.

as this was causing issues for non-threaded workloads.

But there's cases where the current simple logic is wrong. Specifically,
if the pid was a leader pid and the check races with __unhash_process().
Stabilize this by using the pidfd waitqueue lock.

To: Oleg Nesterov <oleg@redhat.com>
Cc: linux-fsdevel@vger.kernel.org
Cc: Jeff Layton <jlayton@kernel.org>
Cc: Lennart Poettering <lennart@poettering.net>
Cc: Daan De Meyer <daan.j.demeyer@gmail.com>
Cc: Mike Yuan <me@yhndnzj.com>
Cc: linux-kernel@vger.kernel.org
Cc: Peter Ziljstra <peterz@infradead.org>
Signed-off-by: Christian Brauner <brauner@kernel.org>

--- b4-submit-tracking ---
# This section is used internally by b4 prep for tracking purposes.
{
  "series": {
    "revision": 2,
    "change-id": "20250411-work-pidfs-enoent-747579160c8c",
    "prefixes": [],
    "base-branch": "work.pidfs.fixes"
  }
}