mnt_idmapping: improve fastpaths
We currently always chase a pointer inode->i_sb->s_user_ns whenever we
need to map a uid/gid. Linus reported in [1] that this is noticable
during path lookup.

In the majority of cases we don't need to bother with that pointer chase
because the inode won't be located on a filesystem that's mounted in a
user namespace. The user namespace of the superblock cannot ever change
once it's mounted. So introduce and raise IOP_USERNS on all inodes and
check for that flag in relevant helpers.

[1]: https://lore.kernel.org/CAHk-=whJgRDtxTudTQ9HV8BFw5-bBsu+c8Ouwd_PrPqPB6_KEQ@mail.gmail.com

To: Linus Torvalds <torvalds@linux-foundation.org>
To: linux-fsdevel <linux-fsdevel@vger.kernel.org>
Cc: Mateusz Guzik <mjguzik@gmail.com>
Cc: Al Viro <viro@zeniv.linux.org.uk>
Cc: Jan Kara <jack@suse.cz>
Cc: Jeff Layton <jlayton@kernel.org>
Cc: Josef Bacik <josef@toxicpanda.com>
Signed-off-by: Christian Brauner <brauner@kernel.org>

---
Changes in v2:
- Don't fully inline low-level helpers.
- Link to v1: https://lore.kernel.org/20250416-work-mnt_idmap-s_user_ns-v1-0-273bef3a61ec@kernel.org

--- b4-submit-tracking ---
# This section is used internally by b4 prep for tracking purposes.
{
  "series": {
    "revision": 2,
    "change-id": "20250416-work-mnt_idmap-s_user_ns-eb57ee83e1d6",
    "prefixes": [
      "RFC"
    ],
    "base-branch": "master.v6.15-rc1",
    "history": {
      "v1": [
        "20250416-work-mnt_idmap-s_user_ns-v1-0-273bef3a61ec@kernel.org"
      ]
    }
  }
}