pidfs: support extended attributes
With the incoming support for permanent pidfs dentries we can start
supporting extended attributes on pidfds. This will allow to attach meta
information to tasks. This currently adds support for trusted extended
attributes which is a first natural target.

One natural extension would be to introduce a custom pidfs.* extended
attribute space and allow for the inheritance of extended attributes
across fork() and exec().

The first simple scheme will allow privileged userspace to slap tags
onto pidfds which is useful for e.g., service managers such as systemd.

To: linux-fsdevel@vger.kernel.org
Cc: Jann Horn <jannh@google.com>
Cc: Josef Bacik <josef@toxicpanda.com>
Cc: Jeff Layton <jlayton@kernel.org>
Cc: Daan De Meyer <daan.j.demeyer@gmail.com>
Cc: Lennart Poettering <lennart@poettering.net>
Cc: Mike Yuan <me@yhndnzj.com>
Cc: Zbigniew Jędrzejewski-Szmek <zbyszek@in.waw.pl>
Cc: Alexander Mikhalitsyn <aleksandr.mikhalitsyn@canonical.com>
Signed-off-by: Christian Brauner <brauner@kernel.org>

---
Changes in v2:
- EDITME: describe what is new in this series revision.
- EDITME: use bulletpoints and terse descriptions.
- Link to v1: https://lore.kernel.org/20250617-work-pidfs-xattr-v1-0-d9466a20da2e@kernel.org



--- b4-submit-tracking ---
# This section is used internally by b4 prep for tracking purposes.
{
  "series": {
    "revision": 2,
    "change-id": "20250617-work-pidfs-xattr-1111246fe9b2",
    "prefixes": [
      "RFC"
    ],
    "base-branch": "work.pidfs.base",
    "history": {
      "v1": [
        "20250617-work-pidfs-xattr-v1-0-d9466a20da2e@kernel.org"
      ]
    }
  }
}